Stuff works on my pc. started with board_to_FEN
